Discovered heading is actually int64 under certain circumstances. This was still throwing the parsing off.

+++ src/riderStatus.js
@@ -114,7 +114,7 @@ export default function riderStatus(buffer) {
                         id: 12
                     },
                     heading: {
-                        type: "int32",
+                        type: "int64",
                         id: 13
                     },
                     lean: {
